% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyConvert.R
\name{easyConvert}
\alias{easyConvert}
\title{Convert GeneID to each others}
\usage{
easyConvert(
  species = c("HUMAN", "MUS"),
  queryList,
  queryType = c("ENTREZID", "SYMBOL", "ENSEMBL", "NCBI")
)
}
\arguments{
\item{species}{default is \code{HUMAN},currently only support 2 species,the other is \code{MUS}}

\item{queryList}{the list to be annotated}

\item{queryType}{including \code{ENTREZID},\code{SYMBOL},\code{ENSEMBL}, correspond to the type of your \code{querList}}
}
\description{
use ?easyConvert() fuction to find the details
}
