% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchAlias.R
\name{searchAlias}
\alias{searchAlias}
\title{Search Alias/synonymous name of Genes}
\usage{
searchAlias(species = "HUMAN", querylist)
}
\arguments{
\item{species}{default is \code{HUMAN},currently only support 2 species,the other is \code{MUS}}

\item{queryList}{the list to be annotated}
}
\description{
use ?searchAlias() fuction to find the details
}
